Lecturerlist display reconfigured in the new list component
