+++ src/app/layout/lecturer-page/lecturer-list/lecturer-list.component.ts
@@ -1,4 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { AngularFireDatabase, FirebaseListObservable } from 'angularfire2/database';
+
+import { LecturerService } from './../../../../services/lecturer.service';
+import { Lecturer } from './../../../../models/Lecturer';
 
 @Component({
   selector: 'app-lecturer-list',
@@ -6,9 +10,18 @@ import { Component, OnInit } from '@angular/core';
 })
 export class LecturerListComponent implements OnInit {
 
-  constructor() { }
+  lecturers: FirebaseListObservable<Lecturer[]>;
+  
+  constructor(private db: AngularFireDatabase,
+    private lecturerService: LecturerService) { 
+      this.getAllTheActiveLecturers();
+    }
 
   ngOnInit() {
   }
 
+
+  private getAllTheActiveLecturers() {
+    this.lecturers = this.lecturerService.getAllTheActiveLecturers();
+  }
 }
